move read_xenium in core module and expand API
