+++ docs/source/api.md
@@ -11,6 +11,7 @@
 
     InSituData
     InSituExperiment
+    read_xenium
 ```
 
 ## IO
@@ -22,5 +23,28 @@
 .. autosummary::
     :toctree: generated
 
-    io.read_xenium
+    io.read_qupath_geojson
+    io.write_qupath_geojson
+    io.read_celldata
+    io.read_multicelldata
+    io.read_shapesdata
 ```
+
+## Plotting
+
+```{eval-rst}
+.. module:: insitupy.pl
+.. currentmodule:: insitupy
+
+.. autosummary::
+    :toctree: generated
+
+    pl.plot_spatial
+    pl.plot_cellular_composition
+    pl.cell_abundance_along_axis
+    pl.cell_expression_along_axis
+    pl.volcano_plot
+    pl.plot_colorlegend
+    pl.plot_overview
+```
+
