+++ insitupy/__init__.py
@@ -20,8 +20,8 @@ from . import utils
 from ._core.dataclasses import AnnotationsData, BoundariesData, ImageData
 from ._core.insitudata import InSituData
 from ._core.insituexperiment import InSituExperiment
+from ._core.readers import read_xenium
 from .images.registration import register_images
-from .io.readers import read_xenium
 from .palettes import CustomPalettes
 from .tools.dge import differential_gene_expression
 from .tools.distance import calc_distance_of_cells_from
